+++ src/packages/bOps.Packages.Storage.Linux/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+// Copyright 2026 Fabio Cavallari
+// SPDX-License-Identifier: Apache-2.0
+
+using System.Runtime.CompilerServices;
+using System.Runtime.Versioning;
+
+[assembly: SupportedOSPlatform("linux")]
+[assembly: InternalsVisibleTo("bOps.Packages.Storage.Linux.Tests")]